# docs/openapi.yaml
openapi: 3.0.0
info:
  version: '1.0'
  title: 'Connection endpoints Api '
  description: This  API retrieve all connections for a person
servers:
  # Added by API Auto Mocking Plugin
  - description: SwaggerHub API Auto Mocking
    url: https://virtserver.swaggerhub.com/aykamal2016/UDACONNECT-Person-Api/1.0
  -  description: Person API for udaconnect project
     url: http://localhost:30002/
paths:
  /api/persons:
    get:
      description: Gets the information for all persons
      responses:
        "200":
          description: 'List of person information successfully retrieved '
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Person'
        "400":
          description: 'Invalid input, object invalid'

  /api/persons/{personId}/connection:
    get:
      summary: gets connections by person id
      description: |
        Gets connections by person id
      parameters:
        - name: personId
          in: path
          description:  ID of the person to get near by connections
          required: true
          schema:
            type: integer
        - name: start_date
          in: query
          description:  start date of visit should be combined with 'end_date'
          required: true
          schema:
            type: string
            format: date
        - name: end_date
          in: query
          description:  end date of visit should be combined with 'start_date'
          required: true
          schema:
            type: string
            format: date
        - name: distance
          in: query
          description:  how far is the distance between the person location and his connections
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: search results matching criteria
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Connection'

components:
  schemas:
    Connection:
      type: object
      properties:
        location:
          $ref: '#/components/schemas/Location'
        person:
          $ref: '#/components/schemas/Person'

    Person:
      required:
        - company_name
        - first_name
        - id
        - last_name
      type: object
      properties:
        id:
          type: integer
          example: 1
        first_name:
          type: string
          example: Ayman
        last_name:
          type: string
          example: Moawad
        company_name:
          type: string
          example: BDC

    Location:
      required:
        - id
        - person_id
        - creation_time
        - latitude
        - longtitude
      type: object
      properties:
        id:
          type: integer
          example: 1
        creation_time:
          type: string
          example: '2020-01-05T10:37:06'
        latitude:
          type: number
          example: 35.0585136
        longtitude:
          type: number
          example: -106.57